Update package lock, move function to LoaderUtils
