+++ src/three/renderer/tiles/TilesRenderer.js
@@ -19,7 +19,6 @@ import { ExtendedFrustum } from '../math/ExtendedFrustum.js';
 import { estimateBytesUsed } from '../utils/MemoryUtils.js';
 import { WGS84_ELLIPSOID } from '../math/GeoConstants.js';
 import { GLTFLoader } from 'three/examples/jsm/loaders/GLTFLoader.js';
-import { getWorkingPath } from '../../../core/renderer/utilities/urlExtension.js';
 
 const _mat = new Matrix4();
 const _euler = new Euler();
@@ -571,7 +570,7 @@ export class TilesRenderer extends TilesRendererBase {
 	async parseTile( buffer, tile, extension, uri, abortSignal ) {
 
 		const cached = tile.cached;
-		const workingPath = getWorkingPath( uri );
+		const workingPath = LoaderUtils.getWorkingPath( uri );
 		const fetchOptions = this.fetchOptions;
 
 		const manager = this.manager;
